Application with log trace
In this class we are making new instance of log class, so we can use it
later in the whole project. When you want to add a log trace in your
class, you need to make a new Logger instance for that class.
Example:
Logger log = Logger.getLogger(Application.class); // with this command
you are making new logger instance for that class.
Than you can just send which file are you compiling using log.getName()
method.
There are two most important methods that you will need to use:
log.info(); - for adding some information in log file
log.error() - for adding information about error in log file
All loggers made in the project are writing in the same .txt file.